tests: Move inner test classes to top level
Surefire plugin does not support this for JUnit 4.
Should improve code coverage - dramatically!
