+++ junit4/src/test/java/io/github/cdiunit/junit4/tests/ProducerConfigBaseTest.java
@@ -24,43 +24,18 @@ import jakarta.enterprise.inject.Produces;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
 
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.MethodRule;
-import org.junit.runner.RunWith;
 
 import io.github.cdiunit.AdditionalClasses;
 import io.github.cdiunit.ProducerConfig;
-import io.github.cdiunit.junit4.CdiJUnit;
-import io.github.cdiunit.junit4.CdiRunner;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@AdditionalClasses(TestProducerConfig.Producers.class)
-abstract class TestProducerConfig {
-
-    @RunWith(CdiRunner.class)
-    @TestProducerConfig.ProducerConfigClass(Object.class)
-    @TestProducerConfig.ProducerConfigNum(0)
-    public static class TestWithRunner extends TestProducerConfig {
-
-    }
-
-    @TestProducerConfig.ProducerConfigClass(Object.class)
-    @TestProducerConfig.ProducerConfigNum(0)
-    public static class TestWithRule extends TestProducerConfig {
-
-        @Rule
-        // Use method - not a field - for rules since test class is added to the bean archive.
-        // Weld enforces that no public fields exist in the normal scoped bean class.
-        public MethodRule cdiUnitMethod() {
-            return CdiJUnit.methodRule();
-        }
-
-    }
+@AdditionalClasses(ProducerConfigBaseTest.Producers.class)
+abstract class ProducerConfigBaseTest {
 
     @Inject
     @Named("a")
