+++ junit4/src/test/java/io/github/cdiunit/junit4/tests/TestNamingContextWithRunner.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2024 the original author or authors.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.cdiunit.test.beans;
+package io.github.cdiunit.junit4.tests;
 
-import jakarta.enterprise.inject.Vetoed;
+import org.junit.runner.RunWith;
 
-@Vetoed
-public class ProducedViaMethod {
-    public ProducedViaMethod(int foo) {
+import io.github.cdiunit.junit4.CdiRunner;
 
-    }
+@RunWith(CdiRunner.class)
+public class TestNamingContextWithRunner extends NamingContextBaseTest {
+
 }
